pyproject.toml
[project]
name = "strava-training-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.83.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
]
